+++ test/regress/cover10.sh
@@ -6,9 +6,10 @@ which nvc
 nvc -a $TESTDIR/regress/cover10.vhd -e --cover=all cover10 -r
 
 # Print only covered
-nvc -c --exclude-file $TESTDIR/regress/data/cover10_ef1.txt \
-       --dont-print uncovered,excluded \
-       --report html work/_WORK.COVER10.elab.covdb 2>&1 | tee out.txt
+nvc --cover-report \
+    --exclude-file $TESTDIR/regress/data/cover10_ef1.txt \
+    --dont-print uncovered,excluded \
+    -o html work/_WORK.COVER10.elab.covdb 2>&1 | tee out.txt
 
 # Check nothing uncovered or excluded is there
 if grep SIGNAL_WHICH_IS_UNCOVERED html/hier/*; then
@@ -19,9 +20,10 @@ if grep SIGNAL_WHICH_IS_EXCLUDED html/hier/*; then
 fi
 
 # Print only uncovered
-nvc -c --exclude-file $TESTDIR/regress/data/cover10_ef1.txt \
-       --dont-print covered,excluded \
-       --report html work/_WORK.COVER10.elab.covdb 2>&1 | tee -a out.txt
+nvc --cover-report \
+    --exclude-file $TESTDIR/regress/data/cover10_ef1.txt \
+    --dont-print covered,excluded \
+    -o html work/_WORK.COVER10.elab.covdb 2>&1 | tee -a out.txt
 
 if grep SIGNAL_WHICH_IS_COVERED html/hier/*; then
   exit 1
@@ -31,9 +33,10 @@ if grep SIGNAL_WHICH_IS_EXCLUDED html/hier/*; then
 fi
 
 # Print only excluded
-nvc -c --exclude-file $TESTDIR/regress/data/cover10_ef1.txt \
-       --dont-print covered,uncovered \
-       --report html work/_WORK.COVER10.elab.covdb 2>&1 | tee -a out.txt
+nvc --cover-report \
+    --exclude-file $TESTDIR/regress/data/cover10_ef1.txt \
+    --dont-print covered,uncovered \
+    -o html work/_WORK.COVER10.elab.covdb 2>&1 | tee -a out.txt
 
 if grep SIGNAL_WHICH_IS_COVERED html/hier/*; then
   exit 1