+++ test/regress/cover15.sh
@@ -4,7 +4,7 @@ pwd
 which nvc
 
 nvc --std=2019 -a $TESTDIR/regress/cover15.vhd -e --cover=statement,branch cover15 -r
-nvc -c --report html work/_WORK.COVER15.elab.covdb  2>&1 | tee out.txt
+nvc --cover-report -o html work/_WORK.COVER15.elab.covdb  2>&1 | tee out.txt
 
 # Adjust output to be work directory relative
 sed -i -e "s/[^ ]*regress\/data\//data\//g" out.txt