+++ test/regress/cover20.sh
@@ -9,7 +9,7 @@ mv work/_WORK.COVER20.elab.covdb DB1.covdb
 nvc -a $TESTDIR/regress/cover20.vhd -e -gG_VAL=1 --cover=statement,toggle cover20 -r
 mv work/_WORK.COVER20.elab.covdb DB2.covdb
 
-nvc -c --report html DB1.covdb DB2.covdb 2>&1 | tee out.txt
+nvc --cover-report -o html DB1.covdb DB2.covdb 2>&1 | tee out.txt
 
 # Adjust output to be work directory relative
 sed -i -e "s/[^ ]*regress\/data\//data\//g" out.txt