+++ test/regress/cover9.sh
@@ -5,19 +5,24 @@ which nvc
 
 # Exclude all uncovered things -> Test valid exclude file
 nvc -a $TESTDIR/regress/cover9.vhd -e --cover=all cover9 -r
-nvc -c --exclude-file $TESTDIR/regress/data/cover9_ef1.txt --report html work/_WORK.COVER9.elab.covdb 2>&1 | tee out.txt
+nvc --cover-report --exclude-file $TESTDIR/regress/data/cover9_ef1.txt \
+    -o html work/_WORK.COVER9.elab.covdb 2>&1 | tee out.txt
 
 # Test invalid command in exclude file
-nvc -c --exclude-file $TESTDIR/regress/data/cover9_ef2.txt work/_WORK.COVER9.elab.covdb 2>&1 | tee -a out.txt
+nvc --cover-report --exclude-file $TESTDIR/regress/data/cover9_ef2.txt \
+    -o html1 work/_WORK.COVER9.elab.covdb 2>&1 | tee -a out.txt
 
 # Test invalid bin name
-nvc -c --exclude-file $TESTDIR/regress/data/cover9_ef3.txt work/_WORK.COVER9.elab.covdb 2>&1 | tee -a out.txt
+nvc --cover-report --exclude-file $TESTDIR/regress/data/cover9_ef3.txt \
+    -o html2 work/_WORK.COVER9.elab.covdb 2>&1 | tee -a out.txt
 
 # Bin placed for command
-nvc -c --exclude-file $TESTDIR/regress/data/cover9_ef4.txt work/_WORK.COVER9.elab.covdb 2>&1 | tee -a out.txt
+nvc --cover-report --exclude-file $TESTDIR/regress/data/cover9_ef4.txt \
+    -o html3 work/_WORK.COVER9.elab.covdb 2>&1 | tee -a out.txt
 
 # Hierarchy missing
-nvc -c --exclude-file $TESTDIR/regress/data/cover9_ef5.txt work/_WORK.COVER9.elab.covdb 2>&1 | tee -a out.txt
+nvc --cover-report --exclude-file $TESTDIR/regress/data/cover9_ef5.txt \
+    -o html4 work/_WORK.COVER9.elab.covdb 2>&1 | tee -a out.txt
 
 # Adjust output to be work directory relative
 sed -i -e "s/[^ ]*regress\/data\//data\//g" out.txt